Restructure project folders for websocket and models
Moved websocket logic to internal/websocket, and reorganized model files into pkg/models directory. This change aims to improve the project’s modularity and maintainability by clearly separating concerns and making the structure more intuitive.
