+++ internal/websocket/models/stream.go
@@ -1,4 +1,4 @@
-package deribit
+package models
 
 import (
 	"context"
